restructure session API
