+++ frontend/peerprep/services/sessionService.ts
@@ -33,48 +33,8 @@ const initializeSocket = async () => {
 
 export const socket = initializeSocket();
 
-export const checkUserMatchStatus = async () => {
-  try {
-    const response = await fetch(
-      `${NEXT_PUBLIC_COLLAB_SERVICE_URL}/api/session/check`,
-      {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${await getAccessToken()}`,
-        },
-      },
-    );
+export const isSocketConnected = async () => {
+  const resolvedSocket = await socket;
+  return resolvedSocket ? resolvedSocket.connected : false;
+}
 
-    if (response.status === 200) {
-      return true;
-    }
-    if (response.status === 204) {
-      return false;
-    }
-    throw new Error(`Unexpected status code: ${response.status}`);
-  } catch (error) {
-    throw error;
-  }
-};
-
-export const leaveMatch = async () => {
-  try {
-    const response = await fetch(
-      `${NEXT_PUBLIC_COLLAB_SERVICE_URL}/api/session/leave`,
-      {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${await getAccessToken()}`,
-        },
-      },
-    );
-
-    if (response.status === 200) {
-      return;
-    }
-
-    throw new Error("Unexpected status code");
-  } catch (error) {
-    throw error;
-  }
-};
